+++ src/Schema/Directives/DeleteDirective.php
@@ -10,8 +10,9 @@ use Nuwave\Lighthouse\Support\Contracts\GlobalId;
 use GraphQL\Language\AST\InputValueDefinitionNode;
 use Nuwave\Lighthouse\Exceptions\DirectiveException;
 use Nuwave\Lighthouse\Support\Contracts\FieldResolver;
+use Nuwave\Lighthouse\Support\Contracts\DefinedDirective;
 
-class DeleteDirective extends BaseDirective implements FieldResolver
+class DeleteDirective extends BaseDirective implements FieldResolver, DefinedDirective
 {
     /**
      * The GlobalId resolver.
@@ -39,6 +40,23 @@ class DeleteDirective extends BaseDirective implements FieldResolver
     public function name(): string
     {
         return 'delete';
+    }
+
+    public static function definition(): string
+    {
+        return /* @lang GraphQL */ <<<'SDL'
+"""
+Delete one or more models by their ID.
+The field must have an single non-null argument that may be a list.
+"""
+directive @delete(
+  """
+  Set to `true` to use global ids for finding the model.
+  If set to `false`, regular non-global ids are used.
+  """
+  globalId: Boolean = false
+) on FIELD_DEFINITION
+SDL;
     }
 
     /**