+++ src/Schema/Directives/FirstDirective.php
@@ -7,8 +7,9 @@ use GraphQL\Type\Definition\ResolveInfo;
 use Nuwave\Lighthouse\Schema\Values\FieldValue;
 use Nuwave\Lighthouse\Support\Contracts\FieldResolver;
 use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
+use Nuwave\Lighthouse\Support\Contracts\DefinedDirective;
 
-class FirstDirective extends BaseDirective implements FieldResolver
+class FirstDirective extends BaseDirective implements FieldResolver, DefinedDirective
 {
     /**
      * Name of the directive.
@@ -18,6 +19,27 @@ class FirstDirective extends BaseDirective implements FieldResolver
     public function name(): string
     {
         return 'first';
+    }
+
+    public static function definition(): string
+    {
+        return /* @lang GraphQL */ <<<'SDL'
+"""
+Get the first query result from a collection of Eloquent models.
+"""
+directive @first(  
+  """
+  Specify the class name of the model to use.
+  This is only needed when the default model resolution does not work.
+  """
+  model: String
+
+  """
+  Apply scopes to the underlying query.
+  """
+  scopes: [String!]
+) on FIELD_DEFINITION
+SDL;
     }
 
     /**