+++ src/Schema/Directives/RulesForArrayDirective.php
@@ -5,11 +5,12 @@ namespace Nuwave\Lighthouse\Schema\Directives;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Collection;
 use Nuwave\Lighthouse\Support\Contracts\ProvidesRules;
+use Nuwave\Lighthouse\Support\Contracts\DefinedDirective;
 use Nuwave\Lighthouse\Support\Contracts\ArgDirectiveForArray;
 use Nuwave\Lighthouse\Support\Traits\HasArgumentPath as HasArgumentPathTrait;
 use Nuwave\Lighthouse\Support\Contracts\HasArgumentPath as HasArgumentPathContract;
 
-class RulesForArrayDirective extends BaseDirective implements ArgDirectiveForArray, ProvidesRules, HasArgumentPathContract
+class RulesForArrayDirective extends BaseDirective implements ArgDirectiveForArray, ProvidesRules, HasArgumentPathContract, DefinedDirective
 {
     use HasArgumentPathTrait;
 
@@ -21,6 +22,30 @@ class RulesForArrayDirective extends BaseDirective implements ArgDirectiveForArr
     public function name(): string
     {
         return 'rulesForArray';
+    }
+
+    public static function definition(): string
+    {
+        return /* @lang GraphQL */ <<<'SDL'
+"""
+Run validation on an array itself, using [Laravel built-in validation](https://laravel.com/docs/validation).
+"""
+directive @rulesForArray(
+  """
+  Specify the validation rules to apply to the field.
+  This can either be a reference to any of Laravel\'s built-in validation rules: https://laravel.com/docs/validation#available-validation-rules,
+  or the fully qualified class name of a custom validation rule.
+  """
+  apply: [String!]!
+
+  """
+  Specify the messages to return if the validators fail.
+  Specified as an input object that maps rules to messages,
+  e.g. { email: "Must be a valid email", max: "The input was too long" }
+  """
+  messages: [RulesMessageMap!]
+) on ARGUMENT_DEFINITION | INPUT_FIELD_DEFINITION
+SDL;
     }
 
     /**